+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Added `endpoint` (`node_id`, `address`, `location`) to topic writer init stream logs
+* Fixed a bug when sometimes in cases of context cancellation returner error was not `errors.Is(err, context.Canceled)`
 
 ## v3.126.5
 * Invoke `OnWriterReceiveResult` trace callback when topic writer receives ack from server (enables client-side logging for sync write diagnostics)